# apidefinitionsiot.yaml
swagger: "2.0"
info:
  title: "IoT Operations Partner API"
...
  version: '3.1'
  contact:
    name: "Aruba iotops team"
    email: "wei.huang@hpe.com"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /api/v3/radios/{id}:
    get:
      summary: "Get Radio detail info"
      description: "Required scope: context; permission: readonly"
      operationId: "IngestionService_GetRadioDetail"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3GetRadioDetailResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "radioIdType"
        in: "query"
        required: false
        type: "string"
        enum:
        - "RADIO_ID_TYPE_MAC"
      tags:
      - "Infra"
...
  /api/v3/zigbee/devices/{id}:
    get:
      summary: "Get zigbee device data"
      description: "Get zigbee device data including device id and uuid, the difference between v3 and earlier version is that\nv3 api support getting multi device class and associated device id, earlier version only have one device id.\nRequired scope: context; permission: readonly"
      operationId: "DeviceContextService_GetZigbeeDeviceData"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiv3GetZigbeeDeviceDataResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/data:
    put:
      summary: "Update Zigbee device data"
      description: "update Zigbee device data\nRequired scope: context; permission: readwrite"
      operationId: "DeviceContextService_UpdateZigbeeDeviceData2"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiv3UpdateZigbeeDeviceDataResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "updateInfo"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/apiv3DeviceDataUpdateInfo"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
    patch:
      summary: "Update Zigbee device data"
      description: "update Zigbee device data\nRequired scope: context; permission: readwrite"
      operationId: "DeviceContextService_UpdateZigbeeDeviceData"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiv3UpdateZigbeeDeviceDataResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "updateInfo"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/apiv3DeviceDataUpdateInfo"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/deviceclass:
    put:
      summary: "Update Zigbee device class"
      description: "update Zigbee device class\nRequired scope: context; permission: classify"
      operationId: "DeviceContextService_UpdateZigbeeDeviceClass"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiv3UpdateZigbeeDeviceClassResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "deviceClass"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/v3UpdateZigbeeDeviceClassRequestDeviceClassArray"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/deviceid:
    put:
      summary: "Set zigbee device id"
      description: "Set zigbee device id based on hardware id and device class\nRequired scope: context; permission: readwrite"
      operationId: "DeviceContextService_SetZigbeeDeviceId"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3SetZigbeeDeviceIdResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "deviceClassScope"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/v1DeviceClassScopedData"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/leave:
    put:
      summary: "SendZigbeeLeave"
      description: "Unpair the zigbee device\nRequired scope: zigbee; permission: readwrite."
      operationId: "IngestionService_SendZigbeeLeave"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3SendZigbeeLeaveResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/offline-timeout:
    put:
      summary: "Set Zigbee Offline Timeout"
      description: "Specify the minutes of offline timeout.\nThe range is between 1 minutes to 1440 minutes(24h)."
      operationId: "IngestionService_SetZigbeeOfflineTimeoutValue"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3SetZigbeeOfflineTimeoutValueResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "timeoutData"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/v3SetZigbeeOfflineTimeoutValueRequestTimeoutData"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/packets:
    post:
      summary: "Send Zigbee southbound data"
      description: "Send Zigbee southbound data.\nRequired scope: zigbee; permission: readwrite."
      operationId: "IngestionService_SendZigbeeSouthboundData"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3SendZigbeeSouthboundDataResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "southboundData"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/v3SendZigbeeSouthboundDataRequestSouthboundData"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/devices/{id}/timeout:
    put:
      operationId: "IngestionService_SetZigbeeTimeoutValue"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3SetZigbeeTimeoutValueResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "timeoutData"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/v3SetZigbeeTimeoutValueRequestTimeoutData"
      - name: "idType"
        description: " - ID_TYPE_HARDWARE_ID: Id type hardwareId\n - ID_TYPE_DEVICE_ID: Id type deviceId"
        in: "query"
        required: false
        type: "string"
        enum:
        - "ID_TYPE_DEVICE_ID"
      - name: "deviceClass"
        in: "query"
        required: false
        type: "string"
      tags:
      - "Zigbee"
...
  /api/v3/zigbee/radios/{radioMac}:
    get:
      summary: "Get zigbee radio detailed info"
      description: "Required scope: context; permission: readonly."
      operationId: "IngestionService_GetZigbeeRadioInfo"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3GetZigbeeRadioInfoResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "radioMac"
        in: "path"
        required: true
        type: "string"
      tags:
      - "Zigbee"
  /api/v3/zigbee/stream/packets:
    get:
      summary: "Subscribe Zigbee packet stream"
      description: "Subscribe Zigbee packet stream filtered by the matching mechanisms declared in AppBundle."
      operationId: "IngestionService_SubscribeZigbeePacketStream"
      responses:
        "200":
          description: "A successful response.(streaming responses)"
          schema:
            type: "object"
            properties:
              result:
                $ref: "#/definitions/v3SubscribeZigbeePacketStreamResponse"
              error:
                $ref: "#/definitions/rpcStatus"
            title: "Stream result of v3SubscribeZigbeePacketStreamResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      parameters:
      - name: "fields"
        description: "Additional fields in northbound messages. Currently only support `DeviceClass`"
        in: "query"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      tags:
      - "Zigbee"
definitions:
...
  apiv3DeviceDataUpdateInfo:
    type: "object"
    properties:
      deviceInfo:
        $ref: "#/definitions/commonv1DeviceInfo"
      sensor:
        $ref: "#/definitions/v1Sensor"
      suppressed:
        type: "boolean"
      selfReportedLocation:
        $ref: "#/definitions/commonv1Location"
    title: "Device data update info"
...
  apiv3GetZigbeeDeviceDataResponse:
    type: "object"
    properties:
      mac:
        type: "string"
      name:
        type: "string"
      networkAddress:
        type: "integer"
        format: "int64"
      radio:
        type: "string"
      lqi:
        type: "integer"
        format: "int64"
      lastSeen:
        type: "string"
        format: "int64"
      uptime:
        type: "string"
        format: "int64"
      deviceClassScope:
        type: "object"
        additionalProperties:
          $ref: "#/definitions/v1DeviceClassScopedData"
      sensor:
        $ref: "#/definitions/v1Sensor"
      deviceInfo:
        $ref: "#/definitions/commonv1DeviceInfo"
      online:
        type: "boolean"
      offlineTimeoutMinute:
        type: "integer"
        format: "int64"
...
  apiv3UpdateZigbeeDeviceClassResponse:
    type: "object"
  apiv3UpdateZigbeeDeviceDataResponse:
    type: "object"
  commonv1DeviceInfo:
    type: "object"
    properties:
      deviceName:
        type: "string"
        title: "device name specified by application"
      measuredPower:
        type: "integer"
        format: "int32"
        title: "measured power: RSSI at 1 meter"
    title: "Device info specified by applications"
...
  protobufAny:
    type: "object"
    properties:
      '@type':
        type: "string"
    additionalProperties: {}
...
  rpcStatus:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      details:
        type: "array"
        items:
          $ref: "#/definitions/protobufAny"
...
  v1DeviceClassScopedData:
    type: "object"
    properties:
      deviceId:
        type: "string"
        title: "the device ID of the device in the given device class"
    title: "Device class scoped data"
...
  v1Sensor:
    type: "object"
    properties:
      battery:
        type: "number"
        format: "float"
        description: "Battery level in percentage(%).\n   minimum: 0,\n   maximum: 100,\n   example: \"75\"."
      temperatureC:
        type: "number"
        format: "float"
        description: "Temperature in Celsius(°C).\n   minimum: -273.15,\n   example: \"36.5\"."
      humidity:
        type: "number"
        format: "float"
        description: "Relative humidity in percentage(%).\n   minimum: 0,\n   maximum: 100,\n   example: \"45\"."
      voltage:
        type: "number"
        format: "float"
        description: "Voltage in volt(V).\n   minimum: -3.4028235e+38,\n   maximum: 3.4028235e+38,\n   example: \"12.5\"."
      illumination:
        type: "number"
        format: "float"
        description: "Illumination in lux.\n   minimum: 0,\n   maximum: 3.4028235e+38,\n   example: \"100\"."
      motion:
        type: "boolean"
        description: "Indicator if motion is detected.\n   example: \"true\"."
      current:
        type: "number"
        format: "float"
        description: "Level of current in ampere(A).\n   minimum: -3.4028235e+38,\n   maximum: 3.4028235e+38,\n   example: \"5\"."
      co:
        type: "number"
        format: "float"
        description: "Level of carbon monoxide(CO) in parts per million(ppm).\n   minimum: 0,\n   maximum: 1000000,\n   example: \"10\"."
      co2:
        type: "number"
        format: "float"
        description: "Level of carbon dioxide(CO2) in parts per million(ppm).\n   minimum: 0,\n   maximum: 1000000,\n   example: \"20\"."
      voc:
        type: "number"
        format: "float"
        description: "Level of volatile organic compound(VOC) in parts per million(ppm).\n   minimum: 0,\n   maximum: 1000000,\n   example: \"30\"."
      resistance:
        type: "number"
        format: "float"
        description: "Level of resistance in ohm(Ω).\n   minimum: 0,\n   maximum: 3.4028235e+38,\n   example: \"20.5\"."
      pressure:
        type: "number"
        format: "float"
        description: "Level of pressure in Kilopascal(kPa).\n   minimum: 0,\n   maximum: 3.4028235e+38,\n   example: \"30\"."
      distance:
        type: "number"
        format: "float"
        description: "Distance in Meter(m).\n   minimum: 0,\n   maximum: 3.4028235e+38,\n   example: \"30\"."
      capacitance:
        type: "number"
        format: "float"
        description: "Capacitance in Farad(F).\n   minimum: 0,\n   maximum: 3.4028235e+38,\n   example: \"20.5\"."
      alarm:
        type: "array"
        items:
          $ref: "#/definitions/SensorAlarm"
        description: "Set of active Alarms.\nexample: \"[\\\"ALARM_WATER\\\", \\\"ALARM_SMOKE\\\", \\\"ALARM_FIRE\\\"]\"."
      accelerometer:
        $ref: "#/definitions/SensorAccelerometer"
      inputs:
        $ref: "#/definitions/SensorInputs"
      extended:
        type: "array"
        items:
          $ref: "#/definitions/SensorExtended"
        description: "Key value pairs that can be freely defined.\n   example: \"[{\\\"extendFieldKey\\\":\\\"key1\\\",\\\"extendFieldValue\\\":\\\"value1\\\"},{\\\"extendFieldKey\\\":\\\"key2\\\",\\\"extendFieldValue\\\":\\\"value2\\\"}]\"."
    title: "Customized device sensor value reported by applications"
...
  v1ZigbeeE2PC:
    type: "object"
    properties:
      destEndpoint:
        type: "integer"
        format: "int64"
        title: "Destination endpoint"
      sourceEndpoint:
        type: "integer"
        format: "int64"
        title: "Source endpoint"
      profileId:
        type: "integer"
        format: "int64"
        title: "Profile ID"
      clusterId:
        type: "integer"
        format: "int64"
        title: "Cluster ID"
    title: "Zigbee E2PC"
...
  v3GetRadioDetailResponse:
    type: "object"
    properties:
      radio:
        $ref: "#/definitions/v3GetRadioDetailResponseRadio"
  v3GetRadioDetailResponseRadio:
    type: "object"
    properties:
      id:
        type: "string"
      mac:
        type: "string"
        title: "8bytes MAC address of radio.\nWhen a radio is detected, its MAC might not be ready. The MAC will be updated later"
      onboard:
        type: "boolean"
        title: "Whether the radio is onboard(internal) or USB(external)"
      descriptor:
        type: "string"
        title: "A string describes the radio, for example, ONBOARD/NORDIC_ONBOARD/NORDIC_USB/etc\nThese values come from AP XML file and used in daemon for looking up radios\nONBOARD is TI"
      description:
        type: "string"
        title: "A arbitry string describes what the radio is"
      firmware:
        $ref: "#/definitions/v1RadioFirmware"
      ble:
        $ref: "#/definitions/v1RadioBLE"
      zigbee:
        $ref: "#/definitions/v1RadioZigBee"
      stats:
        $ref: "#/definitions/v1Statistics"
...
  v3GetZigbeeRadioInfoResponse:
    type: "object"
    properties:
      device:
        $ref: "#/definitions/apiv3DeviceInfo"
      onboard:
        type: "boolean"
      channel:
        type: "integer"
        format: "int64"
      rssi:
        type: "integer"
        format: "int32"
      deviceNum:
        type: "integer"
        format: "int64"
      txPower:
        type: "integer"
        format: "int32"
      pans:
        type: "array"
        items:
          $ref: "#/definitions/v3ZigbeeRadioPan"
      mac:
        type: "string"
      apIp:
        type: "string"
...
  v3SendZigbeeLeaveResponse:
    type: "object"
  v3SendZigbeeSouthboundDataRequestSouthboundData:
    type: "object"
    properties:
      e2pc:
        $ref: "#/definitions/v1ZigbeeE2PC"
      payload:
        type: "string"
        format: "byte"
      apsAck:
        type: "boolean"
  v3SendZigbeeSouthboundDataResponse:
    type: "object"
...
  v3SetZigbeeDeviceIdResponse:
    type: "object"
  v3SetZigbeeOfflineTimeoutValueRequestTimeoutData:
    type: "object"
    properties:
      timeoutMinute:
        type: "integer"
        format: "int64"
        description: "specifies the timeout minutes to set the zigbee device offline.\n   minimum: 1,\n   maximum: 1440,\n   example: 10."
  v3SetZigbeeOfflineTimeoutValueResponse:
    type: "object"
  v3SetZigbeeTimeoutValueRequestTimeoutData:
    type: "object"
    properties:
      timeout:
        $ref: "#/definitions/SetZigbeeTimeoutValueRequestTimeout"
  v3SetZigbeeTimeoutValueResponse:
    type: "object"
...
  v3SubscribeZigbeePacketStreamResponse:
    type: "object"
    properties:
      mac:
        type: "string"
      apMac:
        type: "string"
      radioMac:
        type: "string"
      payload:
        type: "string"
        format: "byte"
      e2pc:
        $ref: "#/definitions/v1ZigbeeE2PC"
      deviceClasses:
        type: "array"
        items:
          type: "string"
...
  v3UpdateZigbeeDeviceClassRequestDeviceClassArray:
    type: "object"
    properties:
      deviceClass:
        type: "array"
        items:
          type: "string"
...
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    name: "apikey"
    in: "header"
security:
- ApiKeyAuth: []